pyproject.toml
[project]
name = "cvtailor"
version = "0.1.0"
description = "Human-in-the-loop CV tailoring with Streamlit and Gemini"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0.1",
  "streamlit>=1.35.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "jinja2>=3.1.4",
  "google-genai>=1.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]